chore(patient): issue fixing

+++ src/app/patient-view/assets/components/navbar-patients/navbar-patients.component.ts
@@ -14,21 +14,30 @@ export class NavbarPatientsComponent implements OnInit {
   ngOnInit() {
       const route = this.route.snapshot.routeConfig.path.split('/');
       this.nav.nativeElement.childNodes.forEach(function (tmp) {
-         if (tmp.attributes.routerLink.value.split('/')[2] === route[1]) {
+         if (tmp.attributes.name.value === route[1]) {
              tmp.classList.value = 'nav--nav-active';
          }
       });
       this.navMobile.nativeElement.childNodes.forEach(function (tmp) {
           if (tmp.attributes.value.value.split('/')[1] === route[1]) {
-              tmp.selected = true;
+                  tmp.selected = true;
           }
       });
   }
     onRoute(route): void {
-        this.router.navigate([route]);
+      if (route !== ' ') {
+          this.router.navigate([route]);
+      } else {
+          this.onOff();
+      }
     }
     onOff(): void {
-        localStorage.clear();
+        if (confirm('Bei Bestätigung melden Sie sich von Ihrem Profil ab')) {
+            localStorage.clear();
+            this.router.navigate(['home']);
+        } else {
+            this.ngOnInit();
+        }
     }
 
 }
